File structure fixing
