+++ Assignment.Interfaces/IAddress.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Assignment;
+
+public interface IAddress
+{
+    string StreetAddress { get; }
+    string City { get; }
+    string State { get; }
+    string Zip { get; }
+}
