--- Assignment/Assignment/Assignment.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-<PropertyGroup>
-  <TargetFramework>net9.0</TargetFramework>
-</PropertyGroup>
-
-<!-- <ItemGroup>
-  <Compile Remove="Person.cs" />
-</ItemGroup> -->
-
-<ItemGroup>
-  <None Remove="People.csv" />
-</ItemGroup>
-
-<ItemGroup>
-  <Content Include="People.csv">
-    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-  </Content>
-  <Content Include="Person.cs">
-    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-  </Content>
-</ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Assignment.Interfaces\Assignment.Interfaces.csproj" />
-  </ItemGroup>
-
-</Project>
